ci(lint): adopt golangci-lint v2 and clear surfaced findings
The lint job was failing because golangci/golangci-lint-action@v6
with version=latest pulls v1.64.8, which is built with Go 1.24 and
refuses to lint a project whose go.mod targets 1.26.1 (exit code 3,
"Go language version used to build golangci-lint is lower than the
targeted Go version"). Bring the job back to green by upgrading the
plumbing and clearing every issue the new linter surfaces.

CI + config:
- Action bumped to @v8 (the major that loads v2 configs) with
  the linter pinned at v2.11.4 so a future "latest" refresh can't
  drift the schema underneath us. Both setup-go calls go to 1.26
  so the toolchain matches go.mod.
- .golangci.yaml migrated to the v2 schema (top-level version: "2",
  linters.settings, exclusions block).

Findings cleared (each was already enabled in the prior config and
would have surfaced under v1 too once it could run):
- errorlint: switch the four solanapay URL-parse error wrappers
  from "%w: ...: %v" to "%w: ...: %w" so the parse cause stays
  unwrappable through errors.Is/As.
- unconvert: drop no-op oapitypes.UUID() casts (alias of
  uuid.UUID) and an int64() wrapper around pow10's already-int64
  return.
- staticcheck SA9003: the empty redis-error else-if branches in
  CachedStore now log via slog.Warn instead of swallowing
  silently — best-effort cache failures are now visible.
- gosec G115: introduce a typed getInt32 env loader for spread
  bps; remaining int<->uint casts (Solana slot, DB-checked
  positive USDC amount, HTTP status) are nolint-annotated with
  the bounding invariant inline.
- gosec G404: jitter is decorrelation, not a security primitive —
  switch to math/rand/v2 (goroutine-safe, self-seeding) and drop
  the *rand.Rand field from saga.Backoff and outbox.Dispatcher;
  nolint with reason at the call sites.
- gosec G706: slog handlers escape structured values, so log
  injection isn't a real risk on the err field; nolint with
  reason at both warn sites.
- revive: rename saga.SagaKind to saga.Kind to drop the
  package-name stutter; the Run.SagaKind field name is unchanged
  because it mirrors the SQL column.

+++ cmd/selatpayd/api.go
@@ -54,7 +54,7 @@ func runAPI(ctx context.Context, cfg config.Config, logger *slog.Logger) error {
 
 	q := quoter.New(pool, quoter.NewMockProvider(quoter.DefaultMockRates()), cfg.QuoteHMACSecret, quoter.Options{
 		TTL:       cfg.QuoteTTL,
-		SpreadBps: int32(cfg.QuoteSpreadBps),
+		SpreadBps: cfg.QuoteSpreadBps,
 	})
 
 	handler := api.NewRouter(api.Deps{
